represent sampling priority as a byte, so it doesn't lead to the metrics map being populated

+++ internal-api/src/main/java/datadog/trace/bootstrap/instrumentation/api/AgentTracer.java
@@ -342,7 +342,7 @@ public class AgentTracer {
 
     @Override
     public Integer getSamplingPriority() {
-      return PrioritySampling.UNSET;
+      return (int) PrioritySampling.UNSET;
     }
 
     @Override
